feat: delete validator with a confirm button instead of a modal
